test: extract theme propagation tests
